+++ data_gorvernance/library/data/connect.ini
@@ -1,4 +1,9 @@
 [DG_WEB]
-scheme_domain = http://dg02.dg.rcos.nii.ac.jp
+BASE_URL = http://dg02.dg.rcos.nii.ac.jp
+GOVSHEET_PATH = .dg/gov-sheet.json
+METADATA_PATH = .dg/metadata.json
+GOVRUN_INDEX_PATH = .crates/index.json
+
 [GRDM]
 BASE_URL = https://rdm.nii.ac.jp/
+API_V2_BASE_URL = https://api.rdm.nii.ac.jp/v2/
