+++ data_gorvernance/library/utils/config/__init__.py
@@ -5,3 +5,4 @@
 """
 from . import message
 from . import path_config
+from . import connect
