+++ data_gorvernance/library/utils/dg_web/__init__.py
@@ -3,14 +3,3 @@ from .api import Api
 from .data import Data
 from .form import Form
 from .runcrate_form import RunCrateForm
-
-
-SCHEME = 'http'
-DOMAIN = 'dg02.dg.rcos.nii.ac.jp'
-
-
-GOVSHEET_PATH = '.dg/gov-sheet.json'
-METADATA_PATH = '.dg/metadata.json'
-
-# governed run
-GOVRUN_INDEX_PATH = '.crates/index.json'
