+++ data_gorvernance/library/utils/save.py
@@ -14,6 +14,8 @@ from .log import TaskLog
 from .storage_provider import grdm
 from .time import TimeDiff
 from .widgets import Button, MessageBox
+from library.utils.config import connect as con_config
+from typing import Union
 
 
 def all_sync_path(abs_root: str) -> list:
@@ -75,6 +77,7 @@ class TaskSave(TaskLog):
         self.save_form_box.width = 900
         # 確定ボタン
         self._save_submit_button = Button(width=500)
+        self.API_V2_BASE_URL = con_config.get('GRDM', 'API_V2_BASE_URL')
 
     def get_grdm_params(self) -> Tuple[str, str]:
         """ GRDMのトークンとプロジェクトIDを取得するメソッドです。
@@ -87,7 +90,6 @@ class TaskSave(TaskLog):
         token = ""
         project_id = ""
         try:
-            #token, project_id = Input.get_grdm_connection_parameters()
             token, project_id = get_grdm_connection_parameters()
         except UnusableVault as e:
             message = msg_config.get('form', 'no_vault')
@@ -110,7 +112,7 @@ class TaskSave(TaskLog):
             self.log.error(message)
         return token, project_id
 
-    def define_save_form(self, source: str | List[str]) -> None:
+    def define_save_form(self, source: Union[str, List[str]]) -> None:
         """ GRDMに保存するフォームを作成するメソッドです。
 
         Args:
@@ -157,14 +159,14 @@ class TaskSave(TaskLog):
         self.save_form_box.clear()
         msg = msg_config.get('save', 'doing')
         timediff.start()
-        grdmmain = grdm.Main()
+        grdmmain = grdm.Grdm()
 
         try:
             for i, path in enumerate(self._source):
                 self.save_msg_output.update_info(f'{msg} {i+1}/{size}')
                 grdmmain.sync(
                     token=self.token,
-                    api_url=grdm.API_V2_BASE_URL,
+                    api_url= self.API_V2_BASE_URL,
                     project_id=self.project_id,
                     abs_source = path,
                     abs_root=self._abs_root_path
