+++ data_gorvernance/library/utils/storage_provider/grdm/__init__.py
@@ -3,14 +3,6 @@
 プロジェクトメタデータをアップロードしたり、APIに接続してプロジェクトメタデータとプロジェクトメンバーの情報を取得するためです
 """
 
-from external import *
-from main import *
-
-BASE_URL = 'https://rdm.nii.ac.jp/'
-
-# TODO: BASE_URLに統一したい
-SCHEME = 'https'
-DOMAIN = 'rdm.nii.ac.jp'
-
-API_V2_BASE_URL = 'https://api.rdm.nii.ac.jp/v2/'
-API_DOMAIN = 'api.rdm.nii.ac.jp'
+from .external import External
+from library.utils.storage_provider.grdm.grdm import Grdm
+from .metadata import Metadata
